Refactored the LED code and ported HSV functions from FastLED.
This change is fairly big. We have created a LED class to hold the LED
RGB values. This class also comes with HSV conversion functions that are
ported from FastLED. They have figured out a lot of spectrum vs rainbow
conversion math. Also their code is meant for LED.

This change also led to refactoring of the code and simplification. All
LED String related code is now moved into the LEDString module.

As the original Twang code was designed without gamma correction, I have
added an inverse gamma correction for the simulated output to compensate
for that. Otherwise the output in the emulator can end up being way too
dark.