Made the hsv methods to regular module methods for flexibility.
This change allows us to do in place hsv setting with the nhsv_* method
variants. It also allows us to add the hsv converted values to the
current LED values without having to create new LED objects.

This is the solution to the addition problem we ran into during the
twitch stream.